scripts/requirements.txt
pypdf>=5.0
pymupdf>=1.24.0
pillow>=10.0
ImageHash>=4.3.0
